Revert "r_cons_print fix for very large output ##cons (#17343)"
This reverts commit 1c74f3f729598d7456f348f74a52c8df611c92ca.

+++ libr/include/r_cons.h
@@ -447,8 +447,8 @@ typedef struct r_cons_context_t {
 	RConsGrep grep;
 	RStack *cons_stack;
 	char *buffer;
-	size_t buffer_len;
-	size_t buffer_sz;
+	int buffer_len;
+	int buffer_sz;
 
 	bool breaked;
 	RStack *break_stack;
